Refactor to avoid duplication.
